[#5] feature : 인터페이스 생성 및 api 경로 정리
