+++ backend/src/main/java/movies/controller/CrowlingController.java
@@ -2,8 +2,8 @@ package movies.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import movies.config.constants.ApiUrlConstants;
 import movies.service.CrowlingService;
+import movies.service.impl.CrowlingServiceImpl;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.web.bind.annotation.DeleteMapping;
