+++ backend/src/test/java/movies/controller/AccountControllerTest.java
@@ -1,18 +1,13 @@
 package movies.controller;
 
 import movies.dto.AccountRequestDto;
-import movies.service.AccountService;
+import movies.service.impl.AccountServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 @ExtendWith(MockitoExtension.class)
@@ -22,7 +17,7 @@ public class AccountControllerTest {
     private AccountController accountController;
 
     @Mock
-    private AccountService accountService;
+    private AccountServiceImpl accountService;
 
     private MockMvc mockMvc;
